requirements.txt
Django==2.2.7
django-environ==0.4.5
djangorestframework==3.11.0
django-cors-headers==3.2.0
markdown==3.1.1
django-filter==2.2.0
pytz==2019.3
setuptools==42.0.2
mysqlclient==1.4.6
sqlparse==0.3.0
pyjwt==1.7.1
pandas==0.25.3
xlrd==1.2.0
Pillow==6.2.1
cachetools==3.1.1
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
rsa==4.0
LunarCalendar==0.0.9
ephem==3.7.7.0
prompt_toolkit==3.0.2
python-crontab==2.4.0
gunicorn==20.0.4